requirements.txt
Django==2.1.3
pytz==2019.1
django-filter==2.1.0
djangorestframework==3.9.4
# six==1.12.0
